+++ backend/requirements.txt
@@ -5,3 +5,4 @@ django-cors-headers==3.13.0
 django-debug-toolbar==3.7.0
 Pillow==9.3.0
 sqlparse==0.4.3
+djangorestframework==3.14.0
