+++ backend/zelenyBazar/urls.py
@@ -1,12 +1,17 @@
 from django.urls import path
 from .views import *
+from .views_old import *
+
+# urlpatterns = [
+#     path('listings/', viewListings ),
+#     path('seed/', seedData),
+#     path('listings/<int:id>/', listing),
+#     path('users/', viewUsers),
+#     path('users/<int:id>/', viewUser),
+#     path('comment/<int:id>/', commentDelete),
+#     path('rating/<int:id>/', ratingDelete)
+# ]
 
 urlpatterns = [
-    path('listings/', viewListings ),
-    path('seed/', seedData),
-    path('listings/<int:id>/', listing),
-    path('users/', viewUsers),
-    path('users/<int:id>/', viewUser),
-    path('comment/<int:id>/', commentDelete),
-    path('rating/<int:id>/', ratingDelete)
+
 ]
